Refactor mongodb-datastore url concat, move events function of api into apiUtils
