+++ pkg/api/utils/eventUtils.go
@@ -3,6 +3,7 @@ package api
 import (
 	"encoding/json"
 	"io/ioutil"
+	"log"
 	"net/http"
 	"net/url"
 	"strings"
@@ -31,21 +32,21 @@ type EventFilter struct {
 
 // NewEventHandler returns a new EventHandler
 func NewEventHandler(baseURL string) *EventHandler {
-	scheme := "http"
 	if strings.Contains(baseURL, "https://") {
 		baseURL = strings.TrimPrefix(baseURL, "https://")
 	} else if strings.Contains(baseURL, "http://") {
 		baseURL = strings.TrimPrefix(baseURL, "http://")
-		scheme = "http"
 	}
 	return &EventHandler{
 		BaseURL:    baseURL,
 		AuthHeader: "",
 		AuthToken:  "",
 		HTTPClient: &http.Client{Transport: getClientTransport()},
-		Scheme:     scheme,
+		Scheme:     "http",
 	}
 }
+
+const mongodbDatastoreServiceBaseUrl = "mongodb-datastore"
 
 // NewAuthenticatedEventHandler returns a new EventHandler that authenticates at the endpoint via the provided token
 func NewAuthenticatedEventHandler(baseURL string, authToken string, authHeader string, httpClient *http.Client, scheme string) *EventHandler {
@@ -56,6 +57,10 @@ func NewAuthenticatedEventHandler(baseURL string, authToken string, authHeader s
 
 	baseURL = strings.TrimPrefix(baseURL, "http://")
 	baseURL = strings.TrimPrefix(baseURL, "https://")
+	baseURL = strings.TrimRight(baseURL, "/")
+	if !strings.HasSuffix(baseURL, mongodbDatastoreServiceBaseUrl) {
+		baseURL += "/" + mongodbDatastoreServiceBaseUrl
+	}
 
 	return &EventHandler{
 		BaseURL:    baseURL,
@@ -82,25 +87,13 @@ func (e *EventHandler) getHTTPClient() *http.Client {
 	return e.HTTPClient
 }
 
-// SendEvent sends an event to Keptn
-func (e *EventHandler) SendEvent(event models.KeptnContextExtendedCE) (*models.EventContext, *models.Error) {
-	bodyStr, err := json.Marshal(event)
-	if err != nil {
-		return nil, buildErrorResponse(err.Error())
-	}
-	return post(e.Scheme+"://"+e.getBaseURL()+"/v1/event", bodyStr, e)
-}
-
 // GetEvents returns all events matching the properties in the passed filter object
 func (e *EventHandler) GetEvents(filter *EventFilter) ([]*models.KeptnContextExtendedCE, *models.Error) {
-	var raw string
-	if e.AuthToken != "" {
-		raw = e.Scheme + "://" + e.getBaseURL() + "/mongodb-datastore/event?"
-	} else {
-		raw = e.Scheme + "://" + e.getBaseURL() + "/event?"
-	}
 
-	u, _ := url.Parse(raw)
+	u, err := url.Parse(e.Scheme + "://" + e.getBaseURL() + "/event?")
+	if err != nil {
+		log.Fatal("error parsing url")
+	}
 
 	query := u.Query()
 
@@ -125,58 +118,10 @@ func (e *EventHandler) GetEvents(filter *EventFilter) ([]*models.KeptnContextExt
 
 	u.RawQuery = query.Encode()
 
-	return e.getEvents(u.String(), e)
+	return e.getEvents(u.String())
 }
 
-// GetEvent returns an event specified by keptnContext and eventType
-//
-// Deprecated: this function is deprecated and should be replaced with the GetEvents function
-func (e *EventHandler) GetEvent(keptnContext string, eventType string) (*models.KeptnContextExtendedCE, *models.Error) {
-	return getEvent(e.Scheme+"://"+e.getBaseURL()+"/v1/event?keptnContext="+keptnContext+"&type="+eventType+"&pageSize=10", e)
-}
-
-func getEvent(uri string, api APIService) (*models.KeptnContextExtendedCE, *models.Error) {
-
-	req, err := http.NewRequest("GET", uri, nil)
-	req.Header.Set("Content-Type", "application/json")
-	addAuthHeader(req, api)
-
-	resp, err := api.getHTTPClient().Do(req)
-	if err != nil {
-		return nil, buildErrorResponse(err.Error())
-	}
-	defer resp.Body.Close()
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, buildErrorResponse(err.Error())
-	}
-
-	if resp.StatusCode >= 200 && resp.StatusCode < 300 {
-
-		if len(body) > 0 {
-			var cloudEvent models.KeptnContextExtendedCE
-			err = json.Unmarshal(body, &cloudEvent)
-			if err != nil {
-				return nil, buildErrorResponse(err.Error())
-			}
-
-			return &cloudEvent, nil
-		}
-
-		return nil, nil
-	}
-
-	var respErr models.Error
-	err = json.Unmarshal(body, &respErr)
-	if err != nil {
-		return nil, buildErrorResponse(err.Error())
-	}
-
-	return nil, &respErr
-}
-
-func (e *EventHandler) getEvents(uri string, api APIService) ([]*models.KeptnContextExtendedCE, *models.Error) {
+func (e *EventHandler) getEvents(uri string) ([]*models.KeptnContextExtendedCE, *models.Error) {
 	events := []*models.KeptnContextExtendedCE{}
 	nextPageKey := ""
